menambahkan fitur upload gambar pada portfolio

+++ controllers/PortfolioController.js
@@ -1,4 +1,6 @@
 import Portfolio from "../modules/PortfolioModule.js"
+import path from 'path'
+import fs from 'fs'
 export const getPortfolios = async(req,res)=>{
     try {
         const portfolios = await Portfolio.find()
@@ -22,27 +24,80 @@ export const getPortfolio = async(req,res)=>{
         })
     }
 }
-export const postPortfolio = async(req,res)=>{
-    const {companyName='', description='', tools=[], github='', deploy='', imgUrl=''} = req.body
-    const portfolio = new Portfolio({companyName, description, tools, github, deploy, imgUrl})
-    try {
-        await portfolio.save()
-        res.status(201).json({
-            message: 'Successful save portfolio'
-        })
-    } catch (error) {
-        res.status(500).json({
+export const postPortfolio = (req,res)=>{
+    if(req.files == null) return res.status(400).json({
+        message: 'No file uploaded'
+    })
+    const file = req.files.file
+    const size = file.data.length
+    const ext = path.extname(file.name)
+    const imgName = file.md5 + ext
+    const imgUrl = `${req.protocol}://${req.get('host')}/images/${imgName}`
+    const allowed = ['.png', '.jpg', 'jpeg']
+    if(!allowed.includes(ext.toLowerCase())) return res.status(422).json({
+        message: 'Invalid extension'
+    })
+    if(size > 5000000) return res.status(422).json({
+        message: 'Image must be less then 5mb'
+    })
+    file.mv(`./public/images/${imgName}`, async(error)=>{
+        if(error) return res.status(500).json({
             message: error.message
         })
-    }
+        const {name='', description='', tools=[], github='', deploy=''} = req.body
+        const portfolio = new Portfolio({name, description, tools, github, deploy, imgName, imgUrl})
+        try {
+            await portfolio.save()
+            res.status(201).json({
+                message: 'Successful save portfolio'
+            })
+        } catch (error) {
+            res.status(500).json({
+                message: error.message
+            })
+        }
+    })
 }
 export const updatePortfolio = async(req, res)=>{
-    try {
-        const portfolio = await Portfolio.findById({_id: req.params.id})
-        if(!portfolio) return res.status(404).json({
-            message: 'Portfolio not found'
+    const portfolio = await Portfolio.findById({_id: req.params.id})
+    if(!portfolio) return res.status(404).json({
+        message: 'Portfolio not found'
+    })
+    let newImageName = portfolio.imgName
+    let newImageUrl = portfolio.imgUrl
+    if(req.files == null){
+        newImageName = portfolio.imgName
+    }else{
+        const file = req.files.file
+        const size = file.data.length
+        const ext = path.extname(file.name)
+        const allowed = ['.png', '.jpg', 'jpeg']
+        newImageName = file.md5 + ext
+        newImageUrl = `${req.protocol}://${req.get('host')}/images/${newImageName}`
+        if(!allowed.includes(ext.toLowerCase())) return res.status(422).json({
+            message: 'Invalid extension'
         })
-        await Portfolio.updateOne({_id: req.params.id}, {$set: req.body})
+        if(size > 5000000) return res.status(422).json({
+            message: 'Image must be less then 5mb'
+        })
+        const filepath = `./public/images/${portfolio.imgName}`
+        fs.unlinkSync(filepath)
+        file.mv(`./public/images/${newImageName}`, (error)=>{
+        if(error) return res.status(500).json({
+            message: error.message
+        })
+        })
+    }
+    try {
+        await Portfolio.updateOne({_id: req.params.id},{
+            name: req.body.name,
+            description: req.body.description,
+            tools: req.body.tools,
+            github : req.body.github,
+            deploy: req.body.deploy,
+            imgName: newImageName,
+            imgUrl: newImageUrl
+        })
         res.status(201).json({
             message: 'Successful update portfolio'
         })
@@ -53,14 +108,20 @@ export const updatePortfolio = async(req, res)=>{
     }
 }
 export const deletePortfolio = async(req,res)=>{
+    const portfolio = await Portfolio.findById({_id: req.params.id})
+    if(!portfolio) return res.status(404).json({
+        message: 'Portfolio not found'
+    })
     try {
+        const filepath = `./public/images/${portfolio.imgName}`
+        fs.unlinkSync(filepath)
         await Portfolio.deleteOne({_id: req.params.id})
         res.status(202).json({
-            message: 'Successful delete portfolio'
+            message: 'Successfuly delete portfolio'
         })
     } catch (error) {
-        res.status(404).json({
-            message: 'Portfolio not found'
+        res.status(500).json({
+            message: error.message
         })
     }
 }
